Updated ui, changed llm model to llama3.2:1b, bug fixes

+++ internal/app/middleware_test.go
@@ -12,7 +12,7 @@ import (
 func TestRateLimiter(t *testing.T) {
 	testCases := []struct {
 		name           string
-		rate          float64
+		rate          int
 		burst         int
 		requestCount  int
 		expectedCodes []int
